requirements.txt
pandas==1.3.2
tweepy
python-dotenv
